Removing cache before downloading music.py

+++ plugins/music.py
@@ -50,15 +50,14 @@ async def download_and_send_song(_: Client, message: Message):
 
     try:
         with yt_dlp.YoutubeDL(YDL_OPTIONS) as ydl:
+            ydl.cache.remove()
+
             info_dict = ydl.extract_info(link, download=False)
             audio_file = ydl.prepare_filename(info_dict)
             ydl.process_info(info_dict)
-        rep = '''
-<b>🎧 Uploader @ShadowRazea
-This bot is uploaded on my \
-<a href='https://github.com/shadowrezi/just-userbot'>GitHub</a>
-</b>
-'''.strip()
+
+
+        rep = '<b>🎧 Uploader @ShadowRazea</b>'
         secmul, dur, dur_arr = 1, 0, duration.split(':')
         
         for i in range(len(dur_arr) - 1, -1, -1):
@@ -86,8 +85,10 @@ This bot is uploaded on my \
         print(e)
 
     finally:
-        os.remove(thumb_name)
-        os.remove(audio_file)
+        if os.path.isfile(thumb_name):
+            os.remove(thumb_name)
+        if os.path.isfile(audio_file):
+            os.remove(audio_file)
 
     print(audio_file)
     print(thumb_name)
